requirements.txt
streamlit>=1.37.0
requests>=2.32.0
openai>=1.40.0
python-dotenv>=1.0.1
httpx>=0.27.0
web3>=7.0.0
